+++ src/Managers/InputManager.h
@@ -46,10 +46,10 @@ class InputManager
       \param bool movePlayer - True if the player should move, or False if the world should instead.
       \param bool changeSprite - True if the player's Sprite frame should change, or False if not.
       \param Vector3D &displacement - Number of pixels to move the player/world by. Note: Pass by reference so the function can change it.
-      \param int duration - How long should the movement take (milliseconds).
+      \param float duration - How long should the movement take (milliseconds).
       \return 0 if success, or -1 if the currentDirection isn't one of the 4 valid directions.
     **/
-    int processMovement(bool movePlayer, bool changeSprite, Vector3D &displacement, int duration);
+    int processMovement(bool movePlayer, bool changeSprite, Vector3D &displacement, float duration);
 
 
   private:
